+++ docs/src/components/landing-core/Hero.js
