+++ docs/src/components/landing-core/StudioIntro.js
